update: organize components setup
