+++ src/components/MainContent/Hero/Hero.jsx
@@ -1,11 +1,11 @@
-import "./Hero.css";
-
 import { intersectionService } from "../../../services/intersection-service";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faArrowDown } from "@fortawesome/free-solid-svg-icons";
 
-export default function Hero() {
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+
+import "./Hero.css";
+
+function Hero() {
   const ref = intersectionService.useIntersectionObserver(1, "typed-out");
 
   return (
@@ -28,3 +28,5 @@ export default function Hero() {
     </header>
   );
 }
+
+export { Hero }
