+++ src/components/MainContent/MainContent.jsx
@@ -4,9 +4,10 @@ import {
   skillService,
   socialMediaService,
 } from "../../services/db-service";
+
 import { About } from "./About/About";
 import { Contact } from "./Contact/Contact";
-import Hero from "./Hero/Hero";
+import {Hero} from "./Hero/Hero";
 import { Projects } from "./Projects/Projects";
 import { Skills } from "./Skills/Skills";
