+++ src/components/Navbar/Navbar.jsx
@@ -1,8 +1,8 @@
-import "./Navbar.css";
-
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-export default function Navbar({ element, links, icon = null }) {
+import "./Navbar.css";
+
+function Navbar({ element, links, icon = null }) {
   const linkItems = links.map((link) => (
     <a key={`${link.name}-link`} className={`${element}__link`} href={link.url}>
       {link.name}
@@ -12,3 +12,5 @@ export default function Navbar({ element, links, icon = null }) {
 
   return <nav className={`${element}__nav nav`}>{linkItems}</nav>;
 }
+
+export { Navbar }
